classBox.java: addConnection() so we actually know what nodes are connected to other nodes, includes an automatic resizer. printConnections() prints the connections array for testing purposes.. Possible Getter functions to come.
relationship.java:
setRelationship(.,.) has been modified so that its behavior isn't
doubled in the event that you create a relationship with ones self.
bool isDashed changed to isdashed to avoid double naming with the method
by the same name

textClassBox.java (which should be TestClassBox.java, which actually
isn't a good title anymore anyway):

Added testing for relationships and encoorperating them in with
classBoxes, includes test for resizing connections[] and printing a
complete summary of a relationship.